+++ src/Adapter/InMemory.php
@@ -5,39 +5,41 @@ namespace Innmind\Filesystem\Adapter;
 
 use Innmind\Filesystem\{
     Adapter,
-    Adapter\InMemory\Overwrite,
-    Adapter\InMemory\Merge,
     File,
     Name,
     Directory,
 };
-use Innmind\Immutable\Maybe;
+use Innmind\Immutable\{
+    Maybe,
+    Predicate\Instance,
+};
 
 final class InMemory implements Adapter
 {
     private Directory $root;
-    private Overwrite|Merge $behaviour;
 
-    private function __construct(Overwrite|Merge $behaviour)
+    private function __construct()
     {
         $this->root = Directory::named('root');
-        $this->behaviour = $behaviour;
     }
 
+    /**
+     * @deprecated Use self::emulateFilesystem()
+     */
     public static function new(): self
     {
-        return new self(new Overwrite);
+        return self::emulateFilesystem();
     }
 
     public static function emulateFilesystem(): self
     {
-        return new self(new Merge);
+        return new self;
     }
 
     #[\Override]
     public function add(File|Directory $file): void
     {
-        $this->root = ($this->behaviour)($this->root, $file);
+        $this->root = $this->merge($this->root, $file);
     }
 
     #[\Override]
@@ -62,5 +64,40 @@ final class InMemory implements Adapter
     public function root(): Directory
     {
         return $this->root;
+    }
+
+    private function merge(Directory $parent, File|Directory $file): Directory
+    {
+        if (!$file instanceof Directory) {
+            return $parent->add($file);
+        }
+
+        $file = $parent
+            ->get($file->name())
+            ->keep(Instance::of(Directory::class))
+            ->match(
+                fn($existing) => $this->mergeDirectories($existing, $file),
+                static fn() => $file,
+            );
+
+        return $parent->add($file);
+    }
+
+    private function mergeDirectories(
+        Directory $existing,
+        Directory $new,
+    ): Directory {
+        $existing = $new
+            ->removed()
+            ->filter(static fn($name) => !$new->contains($name))
+            ->reduce(
+                $existing,
+                static fn(Directory $existing, $name) => $existing->remove($name),
+            );
+
+        return $new->reduce(
+            $existing,
+            fn(Directory $directory, $file) => $this->merge($directory, $file),
+        );
     }
 }